lib/Flask/requirement.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cycler==0.10.0
ddtrace==0.49.0
decorator==4.4.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jyserver==0.0.5
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2021.1
pyvi==0.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn-crfsuite==0.3.6
SQLAlchemy==1.4.15
tabulate==0.8.9
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
